+++ booking-telegram-bot/contexts/telegram-bot/application-module/src/services/index.ts
@@ -1,2 +1,1 @@
 export * from './appointment-conversation/index.js'
-export * from './receive-message/index.js'
