--- booking-telegram-bot/contexts/telegram-bot/application-module/src/services/receive-message/index.ts
@@ -1,1 +1,0 @@
-export * from './receive-message.service.js'
